refactor(processor): split render option handling
Move render scale, region, and output-directory validation out of the document processor so the extraction pipeline is easier to inspect independently from filesystem guardrails.
